+++ gltf/src/net/mgsx/gltf/loaders/shared/GLTFLoaderBase.java
@@ -21,6 +21,7 @@ import net.mgsx.gltf.data.camera.GLTFCamera;
 import net.mgsx.gltf.data.extensions.KHRLightsPunctual;
 import net.mgsx.gltf.data.extensions.KHRLightsPunctual.GLTFLight;
 import net.mgsx.gltf.data.extensions.KHRMaterialsIOR;
+import net.mgsx.gltf.data.extensions.KHRMaterialsIridescence;
 import net.mgsx.gltf.data.extensions.KHRMaterialsPBRSpecularGlossiness;
 import net.mgsx.gltf.data.extensions.KHRMaterialsSpecular;
 import net.mgsx.gltf.data.extensions.KHRMaterialsTransmission;
@@ -106,6 +107,7 @@ public class GLTFLoaderBase implements Disposable {
 					}else if(KHRMaterialsVolume.EXT.equals(extension)){
 					}else if(KHRMaterialsIOR.EXT.equals(extension)){
 					}else if(KHRMaterialsSpecular.EXT.equals(extension)){
+					}else if(KHRMaterialsIridescence.EXT.equals(extension)){
 					}else{
 						throw new GLTFUnsupportedException("Extension " + extension + " required but not supported");
 					}